Implement slice assignment
It needs the AT-KEY method to return a rw reference to an array element.

The awkward part of the present solution is that it won't work for
nonexistent keys unless $.autovivify is set. This is awkward,
because if assignment works:

    %doc<name> = "John";

Then the parallel version should also work, regardless of autovivify settings:

    %doc<name1 name2> = <John Doe>;

The more common use case for this is:

    %doc{@keys} = %input{@keys}